# Improve speed of RDKit greedy diverse conformer selection