@SimonBoothroyd 2021-09-27
> (IMO) much more readable

We can maybe agree to disagree here 😉 

> I get a fairly consistent 2x+ speedup from my function

Looking at the timing it seems you typically save about 0.2ms using the code changes assuming I understand the output of `timeit`, which isn't too much.

In general I'd advise caution in investing too much time in micro-optimisations to these functions (same with #1070 where the saving was only microsecond) - bear in mind these are typically coupled with function calls that take on the order of seconds to minutes and so their latency is almost always hidden, and the speed ups your getting would typically require you to run this code thousands of times before you'd notice the speed-up in a real sense.