@lilyminium 2021-09-27
> Looking at the timing it seems you typically save about 0.2ms using the code changes assuming I understand the output of timeit, which isn't too much.

It does decrease but I do think the speed savings will be some factor of total time rather than an absolute difference, because the masked array creation / usage is looped over (masked arrays are notoriously slow). (below timings are per function call)
```
        Shape (3, 3): 0.312 => 0.152 msec (2.1x speedup over 1000 iterations)
      Shape (10, 10): 0.311 => 0.140 msec (2.2x speedup over 1000 iterations)
    Shape (100, 100): 0.324 => 0.146 msec (2.2x speedup over 1000 iterations)
  Shape (1000, 1000): 0.430 => 0.204 msec (2.1x speedup over 1000 iterations)
Shape (10000, 10000): 1.024 => 0.672 msec (1.5x speedup over 500 iterations)
Shape (20000, 20000): 2.115 => 1.335 msec (1.6x speedup over 100 iterations)
```

> In general I'd advise caution in investing too much time in micro-optimisations to these functions (same with #1070 where the saving was only microsecond) - bear in mind these are typically coupled with function calls that take on the order of seconds to minutes and so their latency is almost always hidden, and the speed ups your getting would typically require you to run this code thousands of times before you'd notice the speed-up in a real sense.

Normally I'd agree but these PRs are all code I modified for something else that I thought might be beneficial in the toolkit :-)