@mattwthompson 2022-07-01
It seems this has gone a bit stale, but these changes are probably good and they're definitely small in scope. Could this be rebased onto `main` (previously `topology-biopolymer-refactor`)?